Update Go dependencies.
To have smaller diff when pulling in OpenTelemetry stuff. Also
nice to do it periodically anyway.

The largest change is to Convey assertion library:
  * ShouldEqual now cares about alias types, including "typed" nil.
    - E.g. "user:someone" != identity.Identity("user:someone") now
    - []string(nil) != nil, etc.
  * ShouldEqual can no longer be used to compare pointers
    - There ShouldPointTo now, except it doesn't work with maps
    - Use reflect.ValueOf(...).Pointer() for maps :(
    - ShouldNotEqual hangs in infinite loop if the things being
      compared have reference cycles :(

Other notable changes:
  * gRPC library fixed malformed nil error problem upstream.
  * Starlark has Cmp(...) interface now

R=iannucci@chromium.org

Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4674035
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>

+++ common/data/cmpbin/number_test.go
@@ -110,7 +110,7 @@ func TestRead(t *testing.T) {
 					v, n, err := ReadUint(buf)
 					So(err, ShouldBeNil)
 					So(n, ShouldEqual, len(c.expect))
-					So(v, ShouldEqual, c.val)
+					So(v, ShouldEqual, uint64(c.val))
 				}
 			})
 		}
